+++ packages/desktop/src/renderer/src/util/copyFileInfo.ts
@@ -1,0 +1,7 @@
+export const copyFileName = (pathname: string): void => {
+  window.electron.clipboard.writeText(window.path.basename(pathname))
+}
+
+export const copyFilePath = (pathname: string): void => {
+  window.electron.clipboard.writeText(pathname)
+}